+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 0.0.51
 
 - Explicitly support all menu item attributes: `activebackground`, `activeforeground`, `background`, `bitmap`, `columnbreak`, `compound`, `font`, `foreground`, `hidemargin`, `indicatoron`, `menu`, `offvalue`, `onvalue`, `selectcolor`, `selectimage`, `state`, `underline`, `value`
-
+- Enhanced Hello, Contextual Menu! & Hello, Menu Bar! to display a message box upon selecting a Language/Country
 
 ## 0.0.50